+++ droplet/db/fabfile.py
@@ -9,13 +9,20 @@ env.hosts = ['172.168.2.117']
 env.user = 'root'
 env.password = '123456'
 
-def install():
+def setup_db():
+    if not confirm("Are you sure host " + env.hosts[0] + " is a DB Server?"):
+        abort("Abort!")
     run('apt-get update')
     run('apt-get install -y php5')    
     run('apt-get remove -y apache2')
     run('apt-get install -y nginx')
     run('apt-get install -y php5-curl php5-gd php5-mysql php5-fpm php5-mcrypt zip')
     run('php5enmod mcrypt')
+    run("debconf-set-selections <<< 'mysql-server mysql-server/root_password password root_db_pass'")
+    run("debconf-set-selections <<< 'mysql-server mysql-server/root_password_again password root_db_pass'")
+    run('apt-get install -q -y mysql-server')
+    run("echo -e 'root_db_pass\n n\n Y\n Y\n Y\n Y\n' | /usr/bin/mysql_secure_installation")
+    run('mysql_install_db')
 
     local_php_path = os.getcwd() + '/php.ini'
     remote_php_path = '/etc/php5/fpm/php.ini'
@@ -35,7 +42,7 @@ def install():
     if result.failed and not confirm("Nginx-PHP setup failed. Continue anyway?"):
         abort("Aborting at user request.")    
     
-    run('apt-get install phpmyadmin')
+    #run('apt-get install phpmyadmin')
     run('service php5-fpm restart')
     run('service nginx restart')
     run('timedatectl set-timezone Asia/Jakarta')
